requirements.txt
# ForgeForth Africa - Data Layer Service
# =======================================
# Centralized database microservice for all platform data operations

# Django Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-environ>=0.11.2

# Redis & Caching
django-redis>=5.4.0
redis>=5.0.1

# Async Support
channels>=4.0.0
channels-redis>=4.1.0

# API Documentation
drf-spectacular>=0.27.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.2.0
uvicorn[standard]>=0.27.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-django>=4.7.0

